Feature: Copy Link to Clipboard, FAB Menu, take photos, png icon files.
UI has been further refined and there is now a FAB Menu within GalleryDetailActivity.
Within the menu is a fab that will copy posted file's link to the device's clipboard.
When selecting a picture to upload it is now possible to take a picture directly
from the device's camera.
Vector graphics have been switched out for PNG files.