Feature: Content Provider
Framework for a personal db that will allow the user to see pictures
they have uploaded to Wetfish, allowing them to access all details of the
post along with a deletion link and a link to the video/image uploaded to
Wetfish. This will only be accessible to the user, offline, and within
the user's device.