Feature: Christmas Commit! Gallery Item Details Page, Videos!
Gallery items can now be clicked on to visit a detailed page of that file's
Title, Description, and Tags. Links to the image on Wetfish is also accessible,
with code for the deletion link ready when applicable. UI under construction.

It is also now possible to take and upload videos straight from the app.