Feature: Image Focusing v2, Glide Colors & loading, Snackbar alerts
Clicking on an image will now prompt the user to start an intent to allow
the user to view the file. Should an app be unavailable to view the file,
the user will be informed via Snackbar.

Glide now prioritizes loading files from the local storage and then wetfish
if local storage content has been deleted.

Glide now has a grey background for loading, cyan background for unsupported
file types, and black for a failure to load available content.