Feature: Successfully generate downscaled image
Upon selecting a downscale ratio in the dropdown menu the app will generate
a downscaled image and display that to the user. Upon selecting the other versions
the app will delete previous downscaled images, and upon selecting the original,
will delete all downscaled images. Appropriate snackbars exist to depict the success
of the downscaling process.