Feature: Delete downscaled image when it is not uploaded & Database Addition
Upon the event that the user selects a new resolution or exits the fragment
while a newly created downscaled version of the original file still exists,
the app will now appropriately delete the file upon onDestroy(). Possible
addition to delete and recreate the file during onPause()

The database can a lso now appropriately handle the addition of edited image
locations.