Feature: Current Image resolution/size views and proper back stack functionality
Image Resolution and Image Size views are now present under the image and repopulate
accordingly in response to the user selecting a new resolution option from the spinner.

The backstack when hitting back now properly takes you to the gallery and
removes the previous behaviour of allowing the user to go back to the previous
upload screen.