Feature: Settings UI
Beginning commit of the settings page. This page will allow for the user's
to access a multitude of features, although currently the only feature will be
sorting by asc/desc upload date, desc being the default to reflect the most
recent upload.

Next update will couple the changing of settings to the gallery itself.