Feature: Exif Method to transfer edited EXIF info
Upon a user's FAB click this method will be called to edit the EXIF of the given
image. If the image has already been edited it will simply change the EXIF.  IF
the currently present image is only the original than a new copy of the original
shall be made to keep the integrity of the original file.