Feature: Prevent EXIF Fab running with no selected Checkboxes, add EXIF tags without deletion option
A counter within ExifDataAdapter that is referenced within EditExifFragment
to see if any checkedBoxes currently exist even if a checked box has been triggered
and an EXIF list copy has been instantiated. This prevents the FAB from running for
when no data is present to edit.

Certain EXIF tags are constant and cannot be deleted. These tags have been moved
to the bottom of General EXIF and have had their checkboxes revoked.