Fix & Feature: Reworked Glide Image Loading, Visit Wetfish for images of file system, UI
Glide image loading is now more streamlined, faster and easier to read.

Upon clicking on an image that doesn't exist on the local file system the app
will launch the web browser to view the image on Wetfish if connected to the internet.
If not, the app will inform the user with a snackbar.

TextView added when the file is not on the local file system, hides file size/length/resolution.