Fix: Placeholder is now a white fill of the ImageView
Because of how Glide loads images, the bounds of the PlaceHolder image was
keeping the centerCropTransform from actually filling the whole view. Because
it is only a preview image and can be clicked on for more information, it now
is a white fill of the view to retain an entire fill of the ImageView like before.