Feature: Refactoring of EXIF feature
EXIF data is currently being reworked to occur while a user uploads a file.
This is a beginning commit to mark the progress of this endeavor, transferring
the base methods and logic for this new implementation. This contains methods to
transfer, create, and check to see if EXIF data has been properly transferred during
the first part of the updated handler thread.