Feature: Appropriate application of EXIF changes for videos, uploading, and Alert Dialogs
Codebase changes have been refactored and changed to now account for only
creating an edited EXIF image for jpg/jpeg images and skipping the procedure
for all other file types. The code now properly uploads the files after  the EXIF
procedure if pertinent and will warn the user with an alert dialog if the application
has detected a mistake in the EXIF editing process.