Feature: Additional Export & Import features
Now upon selection of the export option the app will check to see if a previously
existing fileBackup.db exists. If so it will ask the user if they would like to
keep the old fileBackup.db or overwrite it.

Now upon selection of the import option the app will inform the user that it will
be replacing the app's current database with fileBackup.db within the downloads directory.
The app creates a temporary backup of the Wetfish DB just in case something goes wrong
during the method so it can attempt to revert the old DB back.

Snackbars for both importing and exporting now receive specific strings from
import and export methods for detailed information.

Asynchronous Snackbars are present within the AlertDialogs.